fix validation field of Sign Up Account

+++ account/entity/account_entity.go
@@ -6,6 +6,6 @@ type Account struct {
 	Address       string `json:"address"`
 	Email         string `json:"email"`
 	PhoneNumber   string `json:"phoneNumber"`
-	OutletsNumber int64  `json:"outletNumber"`
+	OutletsNumber int    `json:"outletNumber"`
 	BusinessType  int    `json:"businessType"`
 }
